+++ src/main/java/com/qa/gamestore/service/OrdersService.java
@@ -7,34 +7,36 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.qa.gamestore.domain.Accounts;
+import com.qa.gamestore.domain.Games;
+import com.qa.gamestore.domain.OrderGames;
 import com.qa.gamestore.domain.Orders;
 import com.qa.gamestore.exceptions.IdNotFoundException;
-import com.qa.gamestore.repo.AccountsRepo;
+import com.qa.gamestore.repo.OrderGamesRepo;
 import com.qa.gamestore.repo.OrdersRepo;
 
 @Service
 public class OrdersService implements ServiceInterface<Orders> {
-	
+
 	private OrdersRepo repo;
-	private AccountsRepo accRepo;
-	
+	private OrderGamesRepo ordGRepo;
+	private AccountsService aSer;
+	private GamesService gSer;
+
 	@Autowired
-	public OrdersService(OrdersRepo repo, AccountsRepo accRepo) {
+	public OrdersService(OrdersRepo repo, AccountsService aSer, GamesService gSer, OrderGamesRepo ordGRepo) {
 		this.repo = repo;
-		this.accRepo = accRepo;
+		this.aSer = aSer;
+		this.gSer = gSer;
+		this.ordGRepo = ordGRepo;
 	}
-	
+
 	@Override
-	public Orders create(Orders order) { 
-		Optional<Accounts> opAcc = accRepo.findById(order.getAccountsId());
-		if (opAcc.isPresent()) {
-			Accounts existingAcc = opAcc.get();
-			order.setAccounts(existingAcc);
-			return this.repo.save(order);
-		}
-		return null;
+	public Orders create(Orders order) {
+		Accounts account = aSer.readById(order.getAccountsId());
+		order.setAccounts(account);
+		return this.repo.save(order);
 	}
-	
+
 	@Override
 	public List<Orders> readAll() {
 		return this.repo.findAll();
@@ -42,21 +44,22 @@ public class OrdersService implements ServiceInterface<Orders> {
 
 	@Override
 	public Orders readById(Long id) {
-		//TO-DO exception handling
 		Optional<Orders> opOrder = this.repo.findById(id);
-		return opOrder.get();
+		if (opOrder.isPresent()) {
+			return opOrder.get();
+		} else {
+			return null;
+		}
 	}
 
 	@Override
 	public Orders update(Long id, Orders newOrder) {
-		//TO-DO exception handling
-		Optional<Accounts> opAcc = accRepo.findById(newOrder.getAccountsId());
+		Accounts account = aSer.readById(newOrder.getAccountsId());
 		Optional<Orders> opOrder = this.repo.findById(id);
-		if (opOrder.isPresent() & opAcc.isPresent()) {
+		if (opOrder.isPresent()) {
 			Orders existingOrder = opOrder.get();
-			Accounts existingAcc = opAcc.get();
 			existingOrder.updateFields(newOrder);
-			newOrder.setAccounts(existingAcc);
+			newOrder.setAccounts(account);
 			return this.repo.save(existingOrder);
 		}
 		return null;
@@ -65,13 +68,27 @@ public class OrdersService implements ServiceInterface<Orders> {
 	@Override
 	public boolean delete(Long id) {
 		try {
-			this.repo.deleteById(id);
+			if (repo.findById(id).isPresent()) {
+				this.repo.deleteById(id);
+			} else {
+				throw new IdNotFoundException();
+			}
 		} catch (IdNotFoundException e) {
-			//TO-DO deal with more specific exceptions and log them using Loggers
 			return false;
 		}
 		return !(this.repo.existsById(id)); //return true if delete successful
 	}
-	
-	
+
+	// ### custom query methods go below ###
+
+	public OrderGames add(OrderGames orderGame) {
+		Orders order = this.readById(orderGame.getOrdersId());
+		Games game = gSer.readById(orderGame.getGamesId());
+		orderGame.setOrders(order);
+		orderGame.setGames(game);
+		return this.ordGRepo.save(orderGame);
+	}
+
+
+
 }
